+++ src/main/java/ru/tB/command/CommandContainer.java
@@ -2,19 +2,21 @@ package ru.tB.command;
 
 import ru.tB.service.SendBotMessageService;
 import com.google.common.collect.ImmutableMap;
+import ru.tB.service.TelegramUserService;
 
 public class CommandContainer{
 
     private final ImmutableMap<String, Command> commandMap;
     private final Command unknownCommand;
 
-    public CommandContainer(SendBotMessageService sendBotMessageService) {
+    public CommandContainer(SendBotMessageService sendBotMessageService, TelegramUserService telegramUserService) {
 
         commandMap = ImmutableMap.<String, Command>builder()
-                .put(CommandName.START.getCommandName(), new StartCommand(sendBotMessageService))
-                .put(CommandName.STOP.getCommandName(), new StopCommand(sendBotMessageService))
+                .put(CommandName.START.getCommandName(), new StartCommand(sendBotMessageService, telegramUserService))
+                .put(CommandName.STOP.getCommandName(), new StopCommand(sendBotMessageService, telegramUserService))
                 .put(CommandName.HELP.getCommandName(), new HelpCommand(sendBotMessageService))
                 .put(CommandName.NO.getCommandName(), new NoCommand(sendBotMessageService))
+                .put(CommandName.STAT.getCommandName(), new StatCommand(sendBotMessageService, telegramUserService))
                 .build();
 
         unknownCommand = new UnknownCommand(sendBotMessageService);