+++ src/main/java/ru/tB/repository/TelegramUserRepository.java
@@ -1,0 +1,12 @@
+package ru.tB.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.tB.repository.entity.TelegramUser;
+
+import java.util.List;
+
+@Repository
+public interface TelegramUserRepository extends JpaRepository<TelegramUser, String> {
+    List<TelegramUser> findAllByActiveTrue();
+}