+++ src/test/java/ru/tB/command/CommandContainerTest.java
@@ -1,14 +1,11 @@
-package ru.tB;
+package ru.tB.command;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import ru.tB.command.Command;
-import ru.tB.command.CommandContainer;
-import ru.tB.command.CommandName;
-import ru.tB.command.UnknownCommand;
 import ru.tB.service.SendBotMessageService;
+import ru.tB.service.TelegramUserService;
 
 import java.util.Arrays;
 
@@ -19,7 +16,8 @@ public class CommandContainerTest {
     @BeforeEach
     public void init() {
         SendBotMessageService sendBotMessageService = Mockito.mock(SendBotMessageService.class);
-        commandContainer = new CommandContainer(sendBotMessageService);
+        TelegramUserService telegramUserService = Mockito.mock(TelegramUserService.class);
+        commandContainer = new CommandContainer(sendBotMessageService, telegramUserService);
     }
 
     @Test