+++ src/test/java/ru/tB/command/NoCommandTest.java
@@ -1,9 +1,6 @@
-package ru.tB;
+package ru.tB.command;
 
 import org.junit.jupiter.api.DisplayName;
-import ru.tB.command.Command;
-import ru.tB.command.CommandName;
-import ru.tB.command.NoCommand;
 
 @DisplayName("Unit-level testing for NoCommand")
 public class NoCommandTest extends AbstractCommandTest {