+++ RELEASE_NOTES.md
@@ -2,3 +2,4 @@
 
 * add stub telegram bot
 * add SbringBoot skeleton project
+* add repository layer